// src/ai-fabric/types.ts
export type McpServerStatus =
  | "UNKNOWN"
  | "ON_RESOURCE_ALLOCATION"
  | "AVAILABLE"
  | "IMAGE_UNAVAILABLE"
  | "WAITING_FOR_SCRAPPING"
  | "ON_DELETION"
  | "DELETED"
  | "ON_SUSPENDING"
  | "SUSPENDED"
  | "FAILED"
  | "RUNNING"
  | "COOLED";

export type McpTool = {
  name: string;
  description: string;
};

export type McpServer = {
  id: string;
  name: string;
  status: McpServerStatus;
  tools: McpTool[];
  createdAt: string;
};